### 2026-01-19 #9860 @editorialbot
[REVIEW]: APyT: A modular open-source framework for atom probe tomography data evaluation
**Submitting author:** @sebi-85 (<a href="http://orcid.org/0000-0002-8388-3835">Sebastian Eich</a>)
**Repository:** https://github.com/sebi-85/apyt
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.1.0.dev2
