### 2026-01-19 #9860 @lucydot
@sebi-85,

Transferring a comment from the AEiC in the pre-review thread:
```In the meantime, I'll note that the paper is on the longer side for JOSS (we generally target 1000 words or less), so you can take a look and see whether some of the content in the paper might perhaps be better suited for the software documentation.```
