### 2026-01-19 #9860 @lucydot
@openjournals/bcm-eics  - I think this submission is being reviewed against the new criteria, is this correct? As we started the review itself today (and reading the comment from arfon here: https://github.com/openjournals/joss-reviews/issues/8703#issuecomment-3765310157)
