### 2026-02-09 #9860 @mkuehbach
Free-text reviewer comments @mkuehbach

- L22: AP Suite is licensed proprietary software but the policy how to obtain a license for scientific use has become less restricted and the circle of users that are eligible to apply for this has grown.
- L26: The FAU Erlangen-Nürnberg MatlabAtom Probe Toolbox (henceworth AP toolbox) is not an alternative for the proprietary workflow from raw detector, tof data to voltage-and-bowl corrected data.
- L28: Recommend to group all three, the statement made for APAV and libatomprobe, applies also to the AP toolbox, especially to pyccapt (Monajem and Felfer)
the latter of which is maybe better to mention here as it implements these for the OXCART instrument.
- L34 Separation between task types in the source code is not only done in APyT but also APAV and paraprobe-toolbox.
There are TAP-type instrument developments, Inspico, and LEAP-based activities in Stuttgart, how does this contextualize
with the here developed APyT code, does the code support all these instruments, or mainly, only, these or more, or LEAP,
i.e., how cross-cutting is the software. How does the code relate to Inspico, is it a replacement or complementary, or
an unrelated development?
- L49 Data import, python-based this reads like duplicated efforts, www.github.com/atomprobe-tc/ifes_apt_tc_data_modeling should have been mentioned,
explicitly designed oftware for such task, and supported by the IFES.
- L50 Within the context of the German NFDI this is a duplication of efforts, sophisticated activities done by the FAIRmat project should be mentioned
here to put the work properly into perspective. The reviewer  has updated the ifes_apt_tc_data_modeling library and respective pynxtools-apm parsers
to include data artifacts generated from APyT, to acknowledge the addition that APyT yields for users of Stuttgart-designed AP instruments.
- L64: Contextualization for this dataset, what material, from which study, etc.
- Table 1: For completeness it should be mentioned that the Heller et al. paper explicitly mentions that HDF5 files generated by the AP toolbox can be organized
in a directory locally. Agreed, this is might not be an as sophisticated approach as the one taken by APyT but it is at least an option noteworthy.
For full context, it is also important to note that a DB management with searchability at the granularity of individual concepts for ranging, reconstruction,
and raw data acquisition, atom probe domain specific has been implemented already in NOMAD via its pynxtools-apm plugin. That work had been reported already
at APT&M2023. AP Suite is also scriptable (though agreed with limitations), GUI functionalities also offered by APyT optional user interaction pathway.
APTTools, GUI only for some companion apps rangerapp, 3depict.
- L79: Why is working with uncompressed text files a signature of an efficient storage when this is a core design consideration. How is efficiency understood here,
algorithmic, numerical efficient or less specific.
- L80: The hybrid approach chosen is practical, exactly as the one that is used by IVAS/AP Suite
- L85: How fast a query is depends on the number of datasets in the database, hardware, software stack, the complexity and depth of the query, again efficiency
in terms of what, please provide a factual proof.
- L86: Why should metadata for a given analysis workflow be not immutable as well? The formulation should stress which approach was taken to assure that
each analysis runs such that it generates clearly and programmatically distinguishable entities. How does the approach prevent overwriting of previous results?
- L90: Editability of the database using YAML is a practical approach. How are states tracked over time?
- L91: If one works decoupled from a central database, central within which scope?, how is the process of ingesting states back into the db?
Especially relevant in collaborative workflows
- L97: This entire reasoning chain in the paragraph needs to be backed with actual data from a running system, otherwise comparing a file system with a database is questionable as to which performance indicators the argumentation is based on. The raw read and write performance between binary raw files and HDF5 is often much trickier to substantiate and proof so and so faster, slower, or less efficient than it is here stated. Especially, when no details about the hardware are made that was assumed, used, and from
which the database was served. No use case to this sophistication is reported here.
It is correct that incremental updates to HDF5 files can be less efficient but again here the question is as to why at all one should modify artifacts that have been generated along a computational workflow? Modifying existent arrays specifically when these were allocated using specific presets of the library, (contiguous, max dims fixed upon creation), does not have a clear programmatic approach via HDF5 library functionalities. Adding additional arrays is trivial as adding content to binary files, thanks to h5py, which abstracts many of the technical challenges away. S3 storage hasnt been mentioned in the paper here, but is becoming, also in atom probe, a de facto technology that is used, though often with scientists unaware of this or them not making explicit intentional calls via e.g. software like h5pyd.
- L104: The categorization that is here argued that HDF5 seemingly is rather an archival format neglects the substantial number of cases at highest tier HPC centers where HDF5 is one of the workhorses in usage, synchrotrons included. No file format is a silver bullet, and indeed better alternative for specific use cases exist, e.g., NGFF, or are emerging F3. There is no technical reason as to why the experiences can therefore not productively be used also for processing atom probe data. Using text files is conservative, substantiateable if no other visualization tools for content are available but this is all no longer the case, and often such choices are effectively for the reason that scientist at least want to have the feeling that they can modify if needed. It is the reviewer's opinion that such statements do not do their best with educating the atom probe community to improve the situation and motivate the taking of new approaches. In summary, L97-105 is a too optionated paragraph that should be reconsidered.
- L110: generic Wilkinson et al., 2016 cited, while in the meantime an entire
HDF5 data model specific for atom probe has been defined, not only by FAIRmat and the NeXus International Advisory Committee, but also the IFES APT TC by the way, none of these activities are mentioned here, it remains unclear why.
- L111: Future developments summarize practical suggestions and read sound. A thought to consider is that granularizing off development activities and developer time in an already quite small specialists community, where instead working together as a community on fewer but meatier standards and software might be an equally important if not longer lasting
approach.
- L116: The software is well designed, its documentation appropriate, the fact that tests and modern software development practices are applied is a big plus of the paper.

Since the submission of the manuscript to the journal, the requirements on the content and sections
were modified by the JOSS editorial team. These changes affect the present manuscript, see https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future

- The reviewer ran the software following the documentation on an Ubuntu system. The experience was flawless and the precision of the documentation is remarkable, a great job. Non UTF-8 characters in source code is a matter of taste but indeed
a technical challenge to realize unambiguously using HDF5.

- In summary, judged by the technical implementation of the software the manuscript is worth a JOSS publication.
- The clearly missing contextualization of the work within the detailed technical developments that happened within the open-source landscape, how these evolved while APyT was developed, as the github history shows, surplus the fact that publicly funded activities by the German Research Foundation exist that have specifically targeted and urged the community to consolidate activities, including those in the atom probe community, is a weakness of the paper that needs addressing.

- Suggestions where to improve clarity of the argumentation further:
- L23: Limited scripting capabilities, these limits are?
- L25: What is considered a modern data-science framework?
- L77: "auxiliary detection information", "information" != "data"
