### 2026-02-09 #9860 @mkuehbach
Statement on bullet points unticked:

- Collaborative effort: Does the commit history show contributions from multiple developers and evidence of iterative refinement through community feedback?; the history substantiate a single contributor
- State of the field (required section): Do the authors describe how this software compares to other commonly-used packages? If related tools exist, do they provide a clear "build vs. contribute" justification explaining their unique scholarly contribution and why alternatives are insufficient?; several points were made wrt to this in the detailed reviewer comments
- Research Impact Statement (required section): Does the paper include a "Research Impact Statement" section providing evidence of realized impact (publications, external use, integrations) or credible near-term significance (benchmarks, reproducible materials, community-readiness signals)? Is the evidence presented compelling?; this bullet point and the next one I think need to be added, as these werent required by JOSS back then when the manuscript was submitted
- AI usage disclosure (required section): Does the paper include an "AI usage disclosure" section that transparently discloses any use of generative AI in the software creation, documentation, or paper authoring?
- References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?; made references are good, missing contextualization has been mentioned in explicit reviewer comments listed above
