### 2026-02-13 2dd7343
Add option in wrapper for additional peak width scaling
