### 2026-02-15 #9860 @zhangjy-ge
@lucydot I have finished my preliminary review. In addition to my feedback above (https://github.com/openjournals/joss-reviews/issues/9860#issuecomment-3806469401), I have below comments:
1. `collaborative effort`: the author is the only one contributing to the codebase according to commit history
2. `good practice`: overall, most requirements are satisfied. Though the package does not have unit tests, nor it has a quantitative comparisons with existing similar packages. This is something that can be improved.
3. `Automatic tests`: see point 2.
4. `Community guidelines`: : there is no explicit statement, but since the code is hosted on github, usually it is quite intuitive to contribute, report issues, or seek help. But the authors could make this explicit.
5. Missing `AI Usage Disclosure`
6. In `Documentation`, there is no `Statement of Need`, but this is in the software paper. The author may easily copy it to fix this.
7. `Research Impact statement` is missing.

Some other notes:
1. when I ran `apyt_spectrum_aligh`, I received the following warning: `OMP: Info #277: omp_set_nested routine deprecated, please use omp_set_max_active_levels instead`. This may be system dependent, and this is just a note.
2. The example results I got by running the workflow is somehow different from what is presented in the documentation, especially there are some blank results.
