### 2026-02-19 b5ef5a7
Add optional peak width scaling
