### 2026-02-20 6a74889
Update documentation and bump version to 0.1.0.dev3
